# swagger.json
{
    "swagger": "2.0",
    "info": {
        "version": "1.0.0",
        "title": "Swagger Mono-Api"
    },
    "host": "localhost:8081",
    "basePath": "/techmillresource/mono-api/api",
...
    "schemes": [
        "http",
        "https"
    ],
    "securityDefinitions": {
        "uikey": {
            "in": "header",
            "name": "uikey",
            "type": "apiKey"
        },
        "session_key": {
            "in": "header",
            "name": "apisessionkey",
            "type": "apiKey"
        }
    },
    "paths": {
...
        "/changePassword": {
            "post": {
                "tags": [
                    "security"
                ],
                "description": "change password in system",
                "parameters": [{
                    "in": "body",
                    "name": "body",
                    "description": "lender's password to change",
                    "schema": {
                        "$ref": "#/definitions/ChangePassword"
                    }
                }],
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "$ref": "#/definitions/ChangePassword"
                        }
                    },
                    "400": {
                        "description": "Data not found"
                    }
                },
                "security": [{
                    "uikey": [],
                    "session_key": []
                }]
            }
        },
        "/forgotPassword": {
            "post": {
                "tags": [
                    "security"
                ],
                "description": "send profile's password link to email",
                "parameters": [{
                    "in": "body",
                    "name": "body",
                    "description": "profile's email to send password link",
                    "schema": {
                        "$ref": "#/definitions/ForgotPassword"
                    }
                }],
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "$ref": "#/definitions/ForgotPassword"
                        }
                    },
                    "400": {
                        "description": "Data not found"
                    }
                },
                "security": [{
                    "uikey": []
                }]
            }
        },
        "/resetPassword": {
            "post": {
                "tags": [
                    "security"
                ],
                "description": "reset profile's password in system",
                "parameters": [{
                    "in": "body",
                    "name": "body",
                    "description": "profile's password to reset",
                    "schema": {
                        "$ref": "#/definitions/ResetPassword"
                    }
                }],
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "$ref": "#/definitions/ResetPassword"
                        }
                    },
                    "400": {
                        "description": "Data not found"
                    }
                },
                "security": [{
                    "uikey": []
                }]
            }
        },
        "/registration": {
            "post": {
                "tags": [
                    "registration"
                ],
                "description": "register a new profile in system",
                "parameters": [{
                    "in": "body",
                    "name": "body",
                    "description": "profile's data to register",
                    "schema": {
                        "$ref": "#/definitions/Registration"
                    }
                }],
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "$ref": "#/definitions/Registration"
                        }
                    },
                    "400": {
                        "description": "Data not found"
                    }
                },
                "security": [{
                    "uikey": []
                }]
            }
        },
        "/registrationVerification": {
            "post": {
                "tags": [
                    "registration"
                ],
                "description": "verify registred profile in system",
                "parameters": [{
                    "in": "body",
                    "name": "body",
                    "description": "profile to verify",
                    "schema": {
                        "$ref": "#/definitions/RegistrationVerification"
                    }
                }],
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "$ref": "#/definitions/RegistrationVerification"
                        }
                    },
                    "400": {
                        "description": "Data not found"
                    }
                },
                "security": [{
                    "uikey": []
                }]
            }
        },
...
    },
    "definitions": {
...
        "ChangePassword": {
            "type": "object",
            "required": [
                "profileId",
                "oldPassword",
                "newPassword",
                "channel"
            ],
            "properties": {
                "profileId": {
                    "type": "string",
                    "uniqueItems": true
                },
                "oldPassword": {
                    "type": "string",
                    "uniqueItems": true
                },
                "newPassword": {
                    "type": "string",
                    "uniqueItems": true
                },
                "channel": {
                    "type": "string",
                    "uniqueItems": true,
                    "enum": [
                        "mobile",
                        "self service"
                    ]
                }
            }
        },
        "ForgotPassword": {
            "type": "object",
            "required": [
                "loginName",
                "channel"
            ],
            "properties": {
                "loginName": {
                    "type": "string",
                    "uniqueItems": true
                },
                "channel": {
                    "type": "string",
                    "uniqueItems": true,
                    "enum": [
                        "mobile",
                        "self service"
                    ]
                }
            }
        },
        "ResetPassword": {
            "type": "object",
            "required": [
                "password",
                "resetToken",
                "channel"
            ],
            "properties": {
                "password": {
                    "type": "string",
                    "uniqueItems": true
                },
                "resetToken": {
                    "type": "string",
                    "uniqueItems": true
                },
                "channel": {
                    "type": "string",
                    "uniqueItems": true,
                    "enum": [
                        "mobile",
                        "self service"
                    ]
                }
            }
        },
        "Registration": {
            "type": "object",
            "required": [
                "firstName",
                "lastName",
                "emailAddress",
                "phoneNumber",
                "password",
                "channel"
            ],
            "properties": {
                "firstName": {
                    "type": "string",
                    "uniqueItems": true
                },
                "lastName": {
                    "type": "string",
                    "uniqueItems": true
                },
                "emailAddress": {
                    "type": "string",
                    "uniqueItems": true
                },
                "phoneNumber": {
                    "type": "string",
                    "uniqueItems": true
                },
                "password": {
                    "type": "string",
                    "uniqueItems": true
                },
                "channel": {
                    "type": "string",
                    "uniqueItems": true,
                    "enum": [
                        "mobile",
                        "self service"
                    ]
                }
            }
        },
        "RegistrationVerification": {
            "type": "object",
            "required": [
                "token",
                "channel"
            ],
            "properties": {
                "token": {
                    "type": "string",
                    "uniqueItems": true
                },
                "channel": {
                    "type": "string",
                    "uniqueItems": true,
                    "enum": [
                        "mobile",
                        "self service"
                    ]
                }
            }
        },
...
    }
}